pyproject.toml
[project]
name = "jrjModelRegistry"
version = "0.0.16"
description = "True way to save and serve python models"
readme = "README.md"
requires-python = ">=3.8"
authors = [
    { name = "Jason Jafari", email = "me@jasonjafari.com" }
]
license = { text = "MIT" }
keywords = ["machine learning", "model", "saving", "serving"]

classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.12"
]

dependencies = [
    "annotated-types==0.7.0",
    "anyio==4.9.0",
    "boto3==1.38.11",
    "botocore==1.38.11",
    "certifi==2025.4.26",
    "charset-normalizer==3.4.2",
    "click==8.1.8",
    "dill==0.4.0",
    "dnspython==2.7.0",
    "docutils==0.21.2",
    "email_validator==2.2.0",
    "fastapi==0.115.12",
    "fastapi-cli==0.0.7",
    "h11==0.16.0",
    "httpcore==1.0.9",
    "httptools==0.6.4",
    "httpx==0.28.1",
    "id==1.5.0",
    "idna==3.10",
    "iniconfig==2.1.0",
    "jaraco.classes==3.4.0",
    "jaraco.context==6.0.1",
    "jaraco.functools==4.1.0",
    "Jinja2==3.1.6",
    "jmespath==1.0.1",
    "keyring==25.6.0",
    "markdown-it-py==3.0.0",
    "MarkupSafe==3.0.2",
    "mdurl==0.1.2",
    "more-itertools==10.7.0",
    "nh3==0.2.21",
    "packaging==25.0",
    "pluggy==1.5.0",
    "pycryptodomex==3.22.0",
    "pydantic==2.11.4",
    "pydantic_core==2.33.2",
    "Pygments==2.19.1",
    "pymongo==4.12.1",
    "python-dateutil==2.9.0.post0",
    "python-dotenv==1.1.0",
    "python-multipart==0.0.20",
    "PyYAML==6.0.2",
    "pyzipper==0.3.6",
    "readme_renderer==44.0",
    "requests==2.32.3",
    "requests-toolbelt==1.0.0",
    "rfc3986==2.0.0",
    "rich==14.0.0",
    "rich-toolkit==0.14.5",
    "s3transfer==0.12.0",
    "setuptools==80.3.1",
    "shellingham==1.5.4",
    "six==1.17.0",
    "sniffio==1.3.1",
    "starlette==0.46.2",
    "twine==6.1.0",
    "typer==0.15.3",
    "typing-inspection==0.4.0",
    "typing_extensions==4.13.2",
    "urllib3==2.4.0",
    "uvicorn==0.34.2",
    "uvloop==0.21.0",
    "watchfiles==1.0.5",
    "websockets==15.0.1"
]

[project.urls]
Documentation = "https://github.com/JRJSolutions/jrjModelRegistry/blob/main/README.md"
Source = "https://github.com/JRJSolutions/jrjModelRegistry"
Tracker = "https://github.com/JRJSolutions/jrjModelRegistry/issues"

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["jrjModelRegistry*"]
